+++ jeap-spring-boot-jwe-security-it/src/test/java/ch/admin/bit/jeap/jwe/security/it/JweSecurityTestApplication.java
@@ -1,7 +1,5 @@
 package ch.admin.bit.jeap.jwe.security.it;
 
-import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**